Proofs
This document contains proof that the modified algorithm will surely have a solution and at that optimal point, the leader will have a payoff that is less than that provided by the original DOBSS solution but is less prone to follower inconsistencies. The follower payoff with the new algorithm can be greater or less than the follower payoff provided by the Original algorithm.